pyproject.toml
[project]
name = "Deltares-HMS-Docs"
version = "0.1.0"
description = "Documentation utility "
authors = [
    {name = "Mostafa Farrag",email = "mostafa.farrag@deltares.nl"}
]
readme = "README.md"

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

requires-python = ">=3.11"
dependencies = [
    "tqdm (>=4.67.1,<5.0.0)",
    "pyyaml>=6.0.3",
    "pypandoc>=1.16.2",
    "gitpython>=3.1.45",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pre-commit>=4.3.0",
    "pre-commit-hooks>=6.0.0",
    "nbval>=0.11.0",
]

docs = [
    "mkdocs >=1.5.3",
    "mkdocs-material >=9.5.3",
    "mkdocstrings >=0.24.0",
    "mkdocstrings-python >=1.7.5",
    "mike >=2.1.3",
    "mkdocs-jupyter >=0.25.1",
    "mkdocs-autorefs  >=1.2",
    "mkdocs-macros-plugin  >=1.3.7",
    "mkdocs-table-reader-plugin  >=3.1.0",
    "mkdocs-mermaid2-plugin  >=1.2.1",
    "jupyter-contrib-nbextensions  >=0.7.0",
    "mkdocs-panzoom-plugin",
    "notebook<7.0",
    "jupyter",
    "commitizen (>=4.8.3,<5.0.0)",
]


[project.urls]
"Homepage" = "https://github.com/Deltares/toets-svkw"
"Bug Tracker" = "https://github.com/Deltares/toets-svkw/issues"

[project.scripts]
ddocs = "ddocs.markdown:converter"

[tool.hatch.build.targets.wheel]
packages = ["src/ddocs"]
include = ["src/ddocs/data/*"]


[tool.flake8]
ignore = ["E501"]
max-line-length = 120
count = true

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
testpaths = "tests"
norecursedirs = "docs"
markers = [
    "unit: Pure unit tests (fast, isolated, no I/O)(deselect with '-m \"not unit\"')",
    "integration: Tests involving multiple components or lightweight I/O (deselect with '-m \"not integration\"')",
    "e2e: End-to-end tests (CLI, full workflows)(deselect with '-m \"not e2e\"')",
    "slow: Tests that take >1s (deselect with '-m \"not slow\"')",
    "benchmark: Performance benchmarks (deselect with '-m \"not benchmark\"')",
    "property: Hypothesis property-based tests (deselect with '-m \"not property\"')",
    "regression: Regression tests with golden files (deselect with '-m \"not regression\"')",
    "golden: Tests using golden/snapshot files (deselect with '-m \"not golden\"')",
    "mock: marks tests as mock (deselect with '-m \"not mock\"')",
    "long: marks tests as long (deselect with '-m \"not long\"')",
]



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
